+++ app/src/main/java/de/westnordost/streetcomplete/osm/cycleway_separate/SeparateCycleway.kt
@@ -1,10 +1,5 @@
 package de.westnordost.streetcomplete.osm.cycleway_separate
 
-import de.westnordost.streetcomplete.osm.Tags
-import de.westnordost.streetcomplete.osm.cycleway_separate.SeparateCycleway.*
-import de.westnordost.streetcomplete.osm.sidewalk.createSidewalkSides
-import de.westnordost.streetcomplete.osm.sidewalk.hasSidewalk
-
 enum class SeparateCycleway {
     /** No (designated) cycleway */
     NONE,
@@ -19,39 +14,4 @@ enum class SeparateCycleway {
     /** This way is a cycleway only, however it has a sidewalk mapped on the same way, like some
      *  sort of tiny road for cyclists only */
     WITH_SIDEWALK
-}
-
-/** Returns the situation for a separately mapped cycleway */
-fun createSeparateCycleway(tags: Map<String, String>): SeparateCycleway? {
-
-    if (tags["highway"] !in listOf("path", "footway", "bridleway", "cycleway")) return null
-
-    // cycleway implies bicycle=designated
-    val bicycle = tags["bicycle"] ?: if (tags["highway"] == "cycleway") "designated" else null
-    if (bicycle != "designated") return if (bicycle == "yes") ALLOWED else NONE
-
-    if (createSidewalkSides(tags).hasSidewalk()) {
-        return WITH_SIDEWALK
-    }
-
-    // footway implies foot=designated, path implies foot=yes
-    val foot = tags["foot"] ?: when (tags["highway"]) {
-        "footway" -> "designated"
-        "path" -> "yes"
-        else -> null
-    }
-    val horse = tags["horse"] ?: when (tags["highway"]) {
-        "bridleway" -> "designated"
-        "path" -> "yes"
-        else -> null
-    }
-    val yesOrDesignated = listOf("yes", "designated")
-    if (foot !in yesOrDesignated && horse !in yesOrDesignated) return EXCLUSIVE
-
-    val segregated = tags["segregated"] == "yes"
-    return if (segregated) SEGREGATED else NON_SEGREGATED
-}
-
-fun SeparateCycleway.applyTo(tags: Tags) {
-    TODO()
 }
